pyproject.toml
[project]
name = "teal"
description = "PyTorch Experimental Agents Lib (TEAL)"
version = "0.1.0"
dependencies = [
    "torch>=2.0.1",
    "wandb>=0.15.3",
    "hydra-core>=1.3.2",
    "numba>=0.57.0",
    "procgen>=0.10.7",
    "tabulate>=0.9.0",
]

[tool.setuptools.packages]
find = {}
